Decouple server:restart from watch.
